RepoCacheRemoveStatistics::get_bytes_removed return uintmax_t
On 32-bit systems, size_t is not large enough to represent the amount of
freed space if it exceeds 4GiB. We should use a uintmax_t here instead.

+++ include/libdnf5/repo/repo_cache.hpp
@@ -43,10 +43,10 @@ struct LIBDNF_API RepoCacheRemoveStatistics {
     RepoCacheRemoveStatistics(RepoCacheRemoveStatistics && src) noexcept;
     RepoCacheRemoveStatistics & operator=(RepoCacheRemoveStatistics && src) noexcept;
 
-    std::size_t get_files_removed();  // Number of removed files and links.
-    std::size_t get_dirs_removed();   // Number of removed directorires.
-    std::size_t get_bytes_removed();  // Number of removed bytes.
-    std::size_t get_errors();         // Numbes of errors.
+    std::size_t get_files_removed();     // Number of removed files and links.
+    std::size_t get_dirs_removed();      // Number of removed directorires.
+    std::uintmax_t get_bytes_removed();  // Number of removed bytes.
+    std::size_t get_errors();            // Numbes of errors.
 
     RepoCacheRemoveStatistics & operator+=(const RepoCacheRemoveStatistics & rhs) noexcept;
